Add files via upload
In this case, there is both the opportunity to add images and the program to detect the text as well as via the webcam. These files are the result of the content provided by
Nicholas Renotte that was studied and reapplied by me as a tutorial for applying TensorFlow in some cases.